Eliminated many compiler warnings, imrpvoed error handling and reporting in SdifFile reading. Also fixed MacOS line endings on morphtest.py.

+++ Loris/src/AssociateBandwidth.C
@@ -159,7 +159,7 @@ AssociateBandwidth::findRegionBelow( double binfreq )
 		return -1;
 	}
 	else {
-		return std::min( std::floor(binfreq - booger), _surplus.size() - 1. );
+		return int( std::min( std::floor(binfreq - booger), _surplus.size() - 1. ) );
 	}
 }
